requirements.txt
Flask
scikit-learn
Pillow
joblib
